-JimbeApp:
Uncommented code that raises exceptions.
Exceptions are thrown when deleteLocation method is called, and when insertLocation is called with a TaskList!=null. Using Gui application you obtain these exception trying to save a new location with at least a task, or trying to delete a stored location.
Noticed that GetCurrentLocation always return null (take a look at it using wcfTestClient), while getLocations returns corretly the list of locations. StatisticList too is always null (maybe the problem is just restricted to the fact that today no wifi networks are available in my room).